sending to pooltool

+++ src/main/kotlin/com/swiftmako/jormanager/model/QueryTip.kt
@@ -7,9 +7,9 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class QueryTip(
     @Json(name = "blockNo")
-    val blockNo: Int = 0,
+    val blockNo: Long = 0,
     @Json(name = "headerHash")
     val headerHash: String = "",
     @Json(name = "slotNo")
-    val slotNo: Int = 0
+    val slotNo: Long = 0
 )
